nh-search: extract channel validation
