+++ crates/nh-search/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod args;
 
+mod channel;
 mod search;
